Add support for re-exec without loosing IRC connection
This allows you to execute a new version of the bot, while preserving
the connection to the IRC server.

To make this possible, the IRC connect method of Cinch is overridden. If
bot.config.fd is specified, it tries to use the socket for that file
descriptor. Otherwise, the original connect is used.

The environment variable IRC_SERVER_FD is read into bot.config.fd on
process start. After a successful connection, the environment variable
is set to the file descriptor number of the connection.

On the signal HUP, the process will call exec with its own name and
arguments, and pass on the open socket.